requirements.txt
PyGLM == 1.99.0
PyOpenGL == 3.1.5
glfw == 1.11.0
numpy == 1.18.1